requirements.txt
Flask==3.0.3
flask_cors==4.0.1
python-dotenv==1.0.1
supabase==2.5.1
google-generativeai==0.8.2
pandas==2.2.2
Werkzeug==3.0.3
gunicorn
